Add fix for softmax colorization
Note that it appears that the first letter has larger number of top k
swaps.

This might not be so bad, as evidenced in the minmax output, the first
letter (resonating with SpaceByte and our earlier analysis) has a higher
temperature so to speak than the remaining letters in a word.

However, the selection of lower top k for middle values might be worse.
Perhaps this hints at a better decoding scheme where we emphasize the
first syllable of a word, turning off top k for following.