Add Infinite Head Attention Variant
This is a variation of causal self attention that adds the results of
each head (taking a page from JLS Lemma on capacity), and has potential
to improve parameter efficiency -- with particular boost for smaller
models where the head dimension required to be even smaller than the
embedding dimension.

This limit is removed by this variation, and since most transformers are
limited not by the attention by the MLP, allows exploring how we can
potentially do more parallel computation and limit the size, by reducing
the number of layers required to supply the necessary tooken
inter-relations.