Debug and clean logging
Remove reporting of calculation of the training loss for the eval
iterations, this was likely due to some duplicate calculation of the
training loss at each eval iteration.

Clean the formatting for screen logs to make it easier to add more logs,
and to make it more modular.

Add "better than chance" metric, which -- after normalization to estimate
how many of the vocab are actually used -- could provide a way to
measure capabalities of a model agnostic of tokenizer.