Add auto increment max_iter when best is none
We want to have the smallest max_iters possible to show the difference,
for eval speed purposes (though high enough to overcome any initial
learning curves), but we also want to increase if it hits a roadblock.

This should allow the model to get past any noise it encounters by
increating the signal-to-noise-ratio by incrementing the max_iters, and
rerunning the tests.