Add multibar chart to run_experiments_monitor.py
This allows for us to create bar charts for multiple categories, e.g.
optimizer val loss, peak gpu, iteration avg latency, from the
run_experiments monitor.

One can rearrange and sort from the tui before using.

This also introduces a "modal" interface pattern, where we press a
letter to enter a mode.

In this case it is the letter "q", then number of numeric columns to
draw charts, then the number of columns to merge for the plot-label
(useful when a e.g. a single optimizer has multiple modes, and the
combined label is the true chart label).