Refactor pre peri and post ln config
Refactored block variations to be modular, with separate controls for
each stage of the normalization.

Pre is set to true by default
Peri now means directly after mlp or attn (slightly different meaning)
Post is after residual

Note that Peri ln is changed in meaning from both before and after
module, to just directly after -- so that we can separately toggle each
stage.

With our new defaults, if peri ln is enabled, then it will indeed
sandwich the mlp or attn and be true to the definition, however one can
also toggle off the pre-ln and even toggle on post-ln, in which case
peri-ln is not the definition per se but the idea of doing direct
normalization on the output of the mlp or attn.