Add multidataset curriculum demo
This demo prepares for conversation in more of a continuous manner,
instead of abruptly transitioning from one dataset to another, it mixes
the probabilty of sampling from each, interpolating until we are locked
in on the final target.

The hope is that this would prevent catastropic forgetting, and improve
the final model.